+++ categorias/views.py
@@ -1,20 +1,34 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.models import Group
 from django.contrib.auth.decorators import login_required
 from .forms import CategoriaForm
 from .models import Categoria
 
+def is_bibliotecario_ou_admin(user):
+    return (
+        user.is_superuser or
+        user.groups.filter(name='bibliotecario').exists() or
+        user.groups.filter(name='admin').exists()
+    )
+
 @login_required
 def categoria_list(request):
+    if not is_bibliotecario_ou_admin(request.user):
+        return render(request, '404.html', status=404)
     categorias = Categoria.objects.all()
     return render(request, 'categorias/list.html', {'categorias': categorias})
 
 @login_required
 def categoria_detail(request, pk):
+    if not is_bibliotecario_ou_admin(request.user):
+        return render(request, '404.html', status=404)
     categoria = get_object_or_404(Categoria, pk=pk)
     return render(request, 'categorias/detail.html', {'categoria': categoria})
 
 @login_required
 def categoria_create(request):
+    if not is_bibliotecario_ou_admin(request.user):
+        return render(request, '404.html', status=404)
     if request.method == 'POST':
         form = CategoriaForm(request.POST)
         if form.is_valid():
@@ -26,6 +40,8 @@ def categoria_create(request):
 
 @login_required
 def categoria_update(request, pk):
+    if not is_bibliotecario_ou_admin(request.user):
+        return render(request, '404.html', status=404)
     categoria = get_object_or_404(Categoria, pk=pk)
     if request.method == 'POST':
         form = CategoriaForm(request.POST, instance=categoria)
@@ -38,6 +54,8 @@ def categoria_update(request, pk):
 
 @login_required
 def categoria_delete(request, pk):
+    if not is_bibliotecario_ou_admin(request.user):
+        return render(request, '404.html', status=404)
     categoria = get_object_or_404(Categoria, pk=pk)
     if request.method == 'POST':
         categoria.delete()
